nnpkg: prepare for out-of-tree builds
Still assume everything runs from the package root, except the commands,
which can switch to a different directory (and create if it does not exist).
If it needs to create it, it will show this creation and changing.
For our code, we still assume we are in the package root, and restore after the command.