nnpkg: allow out-of-tree builds
To use different directories, we need to tell grep_all where to look for the files
for the generated files (typically makefile), which will be in the builddir
relative to the package root.  For the configure stuff, we will need to re-compute its
path in the package root relative to the build dir however.